Refactor Exceptions and Tasks into packages
