+++ src/main/java/Exceptions/FileNotLoadedException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class FileNotLoadedException extends DukeException {
+    public FileNotLoadedException() {
+        super("No file data. Creating data file and starting new task list...");
+    }
+}
