+++ src/main/java/Exceptions/InvalidCommandException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class InvalidCommandException extends DukeException {
+    public InvalidCommandException() {
+        super("Invalid command entered. See 'help' for a list of commands you can enter");
+    }
+}
