+++ src/main/java/Exceptions/InvalidTaskException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class InvalidTaskException extends DukeException {
+    public InvalidTaskException() {
+        super("Invalid Tasks.Task Id entered, see 'list' for the list of tasks");
+    }
+}
