+++ src/main/java/Exceptions/NoDescriptionException.java
@@ -1,0 +1,7 @@
+package Exceptions;
+
+public class NoDescriptionException extends DukeException {
+    public NoDescriptionException() {
+        super("No description found. Please enter a description after your command.");
+    }
+}
