+++ src/main/java/Tasks/Deadline.java
@@ -1,3 +1,8 @@
+package Tasks;
+
+import Exceptions.DukeException;
+import Tasks.Task;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
