+++ src/main/java/Tasks/Event.java
@@ -1,3 +1,8 @@
+package Tasks;
+
+import Exceptions.DukeException;
+import Tasks.Task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
