+++ src/main/java/Tasks/Todo.java
@@ -1,4 +1,8 @@
-public class Todo extends Task{
+package Tasks;
+
+import Tasks.Task;
+
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
